Allow React pages to document multiple files
Currently, React component documentation pages can only grab props of components in one file. This forces us to write long `index.tsx` files rather than separating the components into multiple files.

This change allows us to provide React Docgen an array of files to parse for each component.
