+++ www/src/pages/components/alert-banner/react/index.mdx
@@ -60,8 +60,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/AlertBanner/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/AlertBanner/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
