+++ www/src/pages/components/alert/react/index.mdx
@@ -102,8 +102,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Alert/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Alert/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
