+++ www/src/pages/components/button-row/react/index.mdx
@@ -69,8 +69,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/ButtonRow/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/ButtonRow/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
