+++ www/src/pages/components/button/react/index.mdx
@@ -197,8 +197,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Button/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Button/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
