+++ www/src/pages/components/carousel/react/index.mdx
@@ -88,8 +88,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Carousel/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Carousel/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
