+++ www/src/pages/components/checkbox/react/index.mdx
@@ -214,8 +214,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Checkbox/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Checkbox/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
