+++ www/src/pages/components/dropdown/react/index.mdx
@@ -165,8 +165,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Dropdown/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Dropdown/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
