+++ www/src/pages/components/fab/react/index.mdx
@@ -55,8 +55,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Fab/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Fab/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
