+++ www/src/pages/components/form-note/react/index.mdx
@@ -53,8 +53,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/FormNote/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/FormNote/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
