+++ www/src/pages/components/grid/react/index.mdx
@@ -126,8 +126,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Grid/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Grid/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
