+++ www/src/pages/components/horizontal-rule/react/index.mdx
@@ -62,8 +62,10 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/HorizontalRule/index.tsx" }
+        reactComponentProps: allFile(
+            filter: {
+                relativePath: { in: ["thumbprint-react/components/HorizontalRule/index.tsx"] }
+            }
         ) {
             ...ReactComponentPropsFragment
         }
