+++ www/src/pages/components/image/react/index.mdx
@@ -148,8 +148,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Image/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Image/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
