+++ www/src/pages/components/input-row/react/index.mdx
@@ -66,8 +66,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/InputRow/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/InputRow/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
