+++ www/src/pages/components/label/react/index.mdx
@@ -76,8 +76,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Label/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Label/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
