+++ www/src/pages/components/link/react/index.mdx
@@ -238,8 +238,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Link/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Link/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
