+++ www/src/pages/components/loader-dots/react/index.mdx
@@ -61,8 +61,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/LoaderDots/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/LoaderDots/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
