+++ www/src/pages/components/modal-base/react/index.mdx
@@ -96,8 +96,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/ModalBase/index.jsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/ModalBase/index.jsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
