+++ www/src/pages/components/modal-curtain/react/index.mdx
@@ -136,8 +136,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/ModalCurtain/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/ModalCurtain/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
