+++ www/src/pages/components/modal-standard/react/index.mdx
@@ -94,8 +94,10 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/ModalStandard/index.jsx" }
+        reactComponentProps: allFile(
+            filter: {
+                relativePath: { in: ["thumbprint-react/components/ModalStandard/index.jsx"] }
+            }
         ) {
             ...ReactComponentPropsFragment
         }
