+++ www/src/pages/components/modal/react/index.mdx
@@ -663,8 +663,15 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Modal/index.tsx" }
+        reactComponentProps: allFile(
+            filter: {
+                relativePath: {
+                    in: [
+                        "thumbprint-react/components/Modal/index.tsx"
+                        "thumbprint-react/components/Modal/ModalAnimatedWrapper/index.tsx"
+                    ]
+                }
+            }
         ) {
             ...ReactComponentPropsFragment
         }
