+++ www/src/pages/components/pill/react/index.mdx
@@ -56,8 +56,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Pill/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Pill/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
