+++ www/src/pages/components/popover/react/index.mdx
@@ -143,8 +143,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Popover/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Popover/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
