+++ www/src/pages/components/radio/react/index.mdx
@@ -194,8 +194,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Radio/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Radio/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
