+++ www/src/pages/components/service-card/react/index.mdx
@@ -104,8 +104,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/ServiceCard/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/ServiceCard/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
