+++ www/src/pages/components/star-rating/react/index.mdx
@@ -101,8 +101,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/StarRating/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/StarRating/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
