+++ www/src/pages/components/text-area/react/index.mdx
@@ -86,8 +86,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/TextArea/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/TextArea/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
