+++ www/src/pages/components/text-input/react/index.mdx
@@ -142,8 +142,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/TextInput/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/TextInput/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
