+++ www/src/pages/components/tooltip/react/index.mdx
@@ -90,8 +90,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Tooltip/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Tooltip/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
