+++ www/src/pages/components/type/react/index.mdx
@@ -112,8 +112,8 @@ export const pageQuery = graphql`
             ...PackageTableFragment
         }
         # Get component props by path to component file.
-        reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Type/index.tsx" }
+        reactComponentProps: allFile(
+            filter: { relativePath: { in: ["thumbprint-react/components/Type/index.tsx"] } }
         ) {
             ...ReactComponentPropsFragment
         }
